Move semsterplan extractor to own file, update UML
This commit moves the semsterplan extractor to an own file and updates the UML model.
Furthermore adds the UML model as an SVG file.
